+++ src/controleur/controleur_salle.php
@@ -21,15 +21,24 @@ function actionGestionSalle($twig,$db) {
     }
 
     if(isset($_GET['idsup'])){
-        $exec=$salle->delete($_GET['idsup']);
-        if (!$exec){
+        $exec1=$salle->deleteById($_GET['idsup']);
+
+        if (!$exec1){
             $form['supprimer'] = false;
-            $form['message'] = 'Problème de suppression dans la table produit';
+            $form['message'] = 'Problème de suppression dans la table optionSalle';
+        }else{
+            $exec=$salle->delete($_GET['idsup']);
+            if (!$exec){
+                $form['supprimer'] = false;
+                $form['message'] = 'Problème de suppression dans la table produit';
+            }
+            else{
+                $form['supprimer'] = true;
+                $form['message'] = 'Salle supprimée avec succès';
+            }
         }
-        else{
-            $form['supprimer'] = true;
-            $form['message'] = 'Salle supprimée avec succès';
-        }
+
+
     }
